+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased changes
 
+* General
+  * BUGFIX: fix `int32` to `ErrorType` mapping. ([#1063](https://github.com/mozilla/glean/pull/1063))
+
 [Full changelog](https://github.com/mozilla/glean/compare/v31.4.0...main)
 
 # v31.4.0 (2020-07-16)